broke editor to smaller pieces in own directory; does not actually upload to storage when not in production
