+++ backend/services/google-cloud.ts
@@ -1,13 +1,21 @@
 import { Storage } from "@google-cloud/storage"
 import * as shortid from "shortid"
 import * as mimeTypes from "mimetypes"
+import { createWriteStream } from "fs"
 
+const isProduction = process.env.NODE_ENV === "production"
 const bucketName = process.env.GOOGLE_CLOUD_STORAGE_BUCKET
 
-const storage = new Storage({
-  projectId: process.env.GOOGLE_CLOUD_STORAGE_PROJECT,
-  keyFilename: process.env.GOOGLE_CLOUD_STORAGE_KEYFILE,
-})
+const storage = isProduction
+  ? new Storage({
+      projectId: process.env.GOOGLE_CLOUD_STORAGE_PROJECT,
+      keyFilename: process.env.GOOGLE_CLOUD_STORAGE_KEYFILE,
+    })
+  : {
+      bucket: () => ({
+        file: () => ({ save: () => {}, delete: () => new Promise(null) }),
+      }),
+    } // heh
 
 const bucket = storage.bucket(bucketName)
 
@@ -16,15 +24,26 @@ export const uploadImage = async ({
   mimeType,
   name = "",
   directory = "",
+  base64 = false,
 }: {
   imageBuffer: Buffer
   mimeType: string
   name?: string
   directory?: string
+  base64?: boolean
 }): Promise<string> => {
+  return Promise.reject(false)
+
   const filename = `${directory ? directory + "/" : ""}${shortid.generate()}${
     name && name !== "" ? "-" + name : ""
   }.${mimeTypes.detectExtension(mimeType)}`
+
+  if (base64) {
+    const base64 = `data:${mimeType};base64,` + imageBuffer.toString("base64")
+
+    return Promise.resolve(base64)
+  }
+
   const file = bucket.file(filename)
   const outputFilename = `https://storage.googleapis.com/${bucketName}/${filename}`
 
@@ -50,6 +69,10 @@ export const uploadImage = async ({
 export const deleteImage = async (filename: string): Promise<boolean> => {
   if (!filename || filename === "") {
     return Promise.resolve(false)
+  }
+
+  if (~filename.indexOf("base64")) {
+    return Promise.resolve(true)
   }
 
   const file = bucket.file(
