+++ frontend/components/Dashboard/CourseEdit/graphql.ts
@@ -1,10 +1,4 @@
-import React, { useState, useEffect } from "react"
-import { Paper } from "@material-ui/core"
-import { createStyles, makeStyles, Theme } from "@material-ui/core/styles"
-import CourseEditForm from "./CourseEditForm"
-import { useMutation, useQuery } from "react-apollo-hooks"
 import { gql } from "apollo-boost"
-import { addImage_addImage as Image } from "./__generated__/addImage"
 
 export const AddCourseMutation = gql`
   mutation addCourse(
@@ -13,6 +7,7 @@ export const AddCourseMutation = gql`
     $photo: ID
     $promote: Boolean
     $start_point: Boolean
+    $hidden: Boolean
     $status: CourseStatus
     $course_translations: [CourseTranslationCreateInput!]
   ) {
@@ -22,6 +17,7 @@ export const AddCourseMutation = gql`
       photo: $photo
       promote: $promote
       start_point: $start_point
+      hidden: $hidden
       status: $status
       course_translations: $course_translations
     ) {
@@ -29,6 +25,9 @@ export const AddCourseMutation = gql`
       slug
       photo {
         id
+        name
+        original
+        original_mimetype
         compressed
         compressed_mimetype
         uncompressed
@@ -53,6 +52,7 @@ export const UpdateCourseMutation = gql`
     $photo: ID
     $promote: Boolean
     $start_point: Boolean
+    $hidden: Boolean
     $status: CourseStatus
     $new_slug: String
     $course_translations: [CourseTranslationWithIdInput!]
@@ -65,6 +65,7 @@ export const UpdateCourseMutation = gql`
       photo: $photo
       promote: $promote
       start_point: $start_point
+      hidden: $hidden
       status: $status
       course_translations: $course_translations
     ) {
@@ -72,6 +73,9 @@ export const UpdateCourseMutation = gql`
       slug
       photo {
         id
+        name
+        original
+        original_mimetype
         compressed
         compressed_mimetype
         uncompressed
@@ -95,9 +99,12 @@ export const CheckSlugQuery = gql`
 `
 
 export const AddImageMutation = gql`
-  mutation addImage($file: Upload!) {
-    addImage(file: $file) {
+  mutation addImage($file: Upload!, $base64: Boolean) {
+    addImage(file: $file, base64: $base64) {
       id
+      name
+      original
+      original_mimetype
       compressed
       compressed_mimetype
       uncompressed
@@ -113,65 +120,3 @@ export const DeleteImageMutation = gql`
     deleteImage(id: $id)
   }
 `
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    title: {
-      textTransform: "uppercase",
-      marginTop: "0.7em",
-      marginBottom: "0.7em",
-    },
-    paper: {
-      padding: "1em",
-    },
-  }),
-)
-
-const CourseEdit = ({ course }: { course: any }) => {
-  const classes = useStyles()
-
-  const addCourse = useMutation(AddCourseMutation)
-  const updateCourse = useMutation(UpdateCourseMutation)
-  const addImage = useMutation(AddImageMutation)
-  const deleteImage = useMutation(DeleteImageMutation)
-  const checkSlug = CheckSlugQuery
-
-  const uploadImage = async (
-    image: File | undefined,
-  ): Promise<Image | null> => {
-    if (!image) {
-      return null
-    }
-
-    const { data, error } = await addImage({ variables: { file: image } })
-
-    console.log(data)
-
-    if (error) {
-      throw new Error("error uploading image: " + error)
-    }
-
-    if (data && data.addImage) {
-      return data.addImage
-    }
-
-    return null
-  }
-
-  return (
-    <section>
-      <Paper elevation={1} className={classes.paper}>
-        <CourseEditForm
-          course={course}
-          checkSlug={checkSlug}
-          addCourse={addCourse}
-          updateCourse={updateCourse}
-          uploadImage={uploadImage}
-          deleteImage={deleteImage}
-        />
-      </Paper>
-    </section>
-  )
-}
-
-export default CourseEdit
