+++ frontend/package.json
@@ -63,6 +63,7 @@
     "module-resolver": "^1.0.0",
     "next": "^9.0.1",
     "next-compose-plugins": "^2.2.0",
+    "next-env": "^1.1.0",
     "next-fonts": "^0.17.0",
     "next-i18next": "^0.44.0",
     "next-optimized-images": "^2.5.3",
